Create stub classes for Watch and Order
Creates initial stub classes for the watch and order models along with classes for controller, service and unit etsts.

As the order will have a @OneToMany relation with the watches in the order, a further entity (WatchOrder) and composite primary key (WatchOrderPK) will be required.

For the Watch model an additional Discount model will be used to specifically handle the discount elements. This is done with a view that the types of potential discounts could be expanded on in the future.